# task-api-w3/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple in-memory CRUD API for managing tasks — built for FlyRank Internship Week 2."
  },
  "paths": {
    "/": {
      "get": {
        "summary": "API info",
        "description": "Returns the API name, version, and available endpoints.",
        "tags": ["Info"],
        "responses": {
          "200": {
            "description": "API metadata",
            "content": {
              "application/json": {
                "example": {
                  "name": "Task API",
                  "version": "1.0",
                  "endpoints": ["/tasks"]
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check",
        "description": "Returns server health status. Real companies ping this endpoint to know a server is alive.",
        "tags": ["Info"],
        "responses": {
          "200": {
            "description": "Server is healthy",
            "content": {
              "application/json": {
                "example": { "status": "ok" }
              }
            }
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "List all tasks",
        "description": "Returns all tasks. Supports optional filtering by done status and keyword search.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "done",
            "in": "query",
            "description": "Filter by completion status (true or false)",
            "required": false,
            "schema": { "type": "boolean" }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Filter tasks whose title contains this keyword (case-insensitive)",
            "required": false,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of tasks",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "title": "Learn Express.js", "done": false },
                  { "id": 2, "title": "Build CRUD API",   "done": true  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new task",
        "description": "Creates a task with the given title. The server assigns the id and sets done to false.",
        "tags": ["Tasks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string", "example": "Buy milk" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created",
            "content": {
              "application/json": {
                "example": { "id": 4, "title": "Buy milk", "done": false }
              }
            }
          },
          "400": {
            "description": "title is missing or empty",
            "content": {
              "application/json": {
                "example": { "error": "title is required and cannot be empty" }
              }
            }
          }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a single task",
        "description": "Returns the task with the given id. Returns 404 if not found.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task id",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "example": { "id": 1, "title": "Learn Express.js", "done": false }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update a task",
        "description": "Updates a task's title and/or done status. Returns the updated task.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task id",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string",  "example": "Updated title" },
                  "done":  { "type": "boolean", "example": true }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "example": { "id": 1, "title": "Updated title", "done": true }
              }
            }
          },
          "400": {
            "description": "Invalid body",
            "content": {
              "application/json": {
                "example": { "error": "title cannot be empty" }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a task",
        "description": "Removes the task from the list. Returns 204 No Content on success.",
        "tags": ["Tasks"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Task id",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "204": { "description": "Task deleted — no body returned" },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "example": { "error": "Task 99 not found" }
              }
            }
          }
        }
      }
    },
    "/stats": {
      "get": {
        "summary": "Task statistics",
        "description": "Returns total, done, and open task counts.",
        "tags": ["Extras"],
        "responses": {
          "200": {
            "description": "Task counts",
            "content": {
              "application/json": {
                "example": { "total": 3, "done": 1, "open": 2 }
              }
            }
          }
        }
      }
    },
    "/reset": {
      "post": {
        "summary": "Reset to seed data",
        "description": "Restores the original 3 example tasks. Useful for demos.",
        "tags": ["Extras"],
        "responses": {
          "200": {
            "description": "Tasks reset",
            "content": {
              "application/json": {
                "example": {
                  "message": "Tasks reset to seed data",
                  "tasks": [
                    { "id": 1, "title": "Learn Express.js", "done": false },
                    { "id": 2, "title": "Build CRUD API",   "done": true  },
                    { "id": 3, "title": "Push to GitHub",   "done": false }
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}